+++ tests/Utils/Rules/EqualFieldRule.php
@@ -10,37 +10,20 @@ class EqualFieldRule implements Rule, WithReferenceRule
     /**
      * @var string
      */
-    protected $field;
-
-    /**
-     * @var string
-     */
-    protected $attribute;
-
-    /**
-     * @var int
-     */
-    protected $reference;
-
-    public function __construct(string $field)
-    {
-        $this->field = $field;
-    }
+    protected $argumentPath;
 
     public function passes($attribute, $value): bool
     {
-        $this->attribute = $attribute;
-
         return false;
     }
 
     public function message(): string
     {
-        return "The $this->attribute must be equal to $this->field.";
+        return $this->argumentPath;
     }
 
     public function setArgumentPath(array $argumentPath): void
     {
-        $this->field = implode('.', array_merge($argumentPath, [$this->field]));
+        $this->argumentPath = implode('.', $argumentPath);
     }
 }