+++ CHANGELOG.md
@@ -13,7 +13,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Added
 
-- Add support for custom rules that reference other fields in the request using the `with_reference` wrapper rule or the `WithReferenceRule` contract for rule classes
+- Add support for custom rules that reference other arguments https://github.com/nuwave/lighthouse/pull/1903
 
 ## v5.16.2